Crate naga_types

Modules

glsl

hlsl

msl

spv

Macros

link_to_wgc_docs

Create a Markdown link definition referring to the wgpu_core crate.

link_to_wgpu_docs

Create a Markdown link definition referring to the wgpu crate.

link_to_wgpu_item

Create a Markdown link definition referring to an item in the wgpu crate.

Structs

ResourceBinding

Pipeline binding information for global resources.

TaskDispatchLimits

Enums

ShaderStage

Stage of the programmable pipeline.

VertexFormat

Corresponds to WebGPU GPUVertexFormat.

Type Aliases

FastHashMap

Hash map that is faster but not resilient to DoS attacks. (Similar to rustc_hash::FxHashMap but using hashbrown::HashMap instead of alloc::collections::HashMap.) To construct a new instance: FastHashMap::default()

FastHashSet

Hash set that is faster but not resilient to DoS attacks. (Similar to rustc_hash::FxHashSet but using hashbrown::HashSet instead of alloc::collections::HashMap.)

FastIndexMap

Insertion-order-preserving hash map (IndexMap<K, V>), but with the same hasher as FastHashMap<K, V> (faster but not resilient to DoS attacks).

FastIndexSet

Insertion-order-preserving hash set (IndexSet<K>), but with the same hasher as FastHashSet<K> (faster but not resilient to DoS attacks).